refactor: organize chat models
