+++ app/src/main/java/org/ole/planet/myplanet/ui/chat/ChatApiHelper.kt
@@ -6,7 +6,7 @@ import okhttp3.RequestBody
 import okhttp3.ResponseBody
 import org.ole.planet.myplanet.datamanager.ApiClient
 import org.ole.planet.myplanet.datamanager.ApiInterface
-import org.ole.planet.myplanet.model.ChatModel
+import org.ole.planet.myplanet.model.chat.ChatModel
 import org.ole.planet.myplanet.utilities.Utilities
 import retrofit2.Call
 import retrofit2.Callback
