+++ app/src/main/java/org/ole/planet/myplanet/ui/chat/ChatDetailFragment.kt
@@ -25,8 +25,17 @@ import org.ole.planet.myplanet.R
 import org.ole.planet.myplanet.databinding.FragmentChatDetailBinding
 import org.ole.planet.myplanet.datamanager.*
 import org.ole.planet.myplanet.ui.chat.ChatApiHelper
-import org.ole.planet.myplanet.model.*
-import org.ole.planet.myplanet.model.RealmChatHistory.Companion.addConversationToChatHistory
+import org.ole.planet.myplanet.model.RealmUserModel
+import org.ole.planet.myplanet.model.chat.ChatRequestModel
+import org.ole.planet.myplanet.model.chat.ContentData
+import org.ole.planet.myplanet.model.chat.AiProvider
+import org.ole.planet.myplanet.model.chat.ChatModel
+import org.ole.planet.myplanet.model.chat.ContinueChatModel
+import org.ole.planet.myplanet.model.chat.Data
+import org.ole.planet.myplanet.model.chat.ChatViewModel
+import org.ole.planet.myplanet.model.chat.RealmChatHistory
+import org.ole.planet.myplanet.model.chat.Conversation
+import org.ole.planet.myplanet.model.chat.RealmChatHistory.Companion.addConversationToChatHistory
 import org.ole.planet.myplanet.service.UserProfileDbHandler
 import org.ole.planet.myplanet.ui.dashboard.DashboardActivity
 import org.ole.planet.myplanet.utilities.Constants
