+++ app/src/main/java/org/ole/planet/myplanet/ui/news/AdapterNews.kt
@@ -30,7 +30,7 @@ import io.realm.RealmList
 import io.realm.Sort
 import org.ole.planet.myplanet.R
 import org.ole.planet.myplanet.databinding.RowNewsBinding
-import org.ole.planet.myplanet.model.Conversation
+import org.ole.planet.myplanet.model.chat.Conversation
 import org.ole.planet.myplanet.model.RealmNews
 import org.ole.planet.myplanet.model.RealmNews.Companion.createNews
 import org.ole.planet.myplanet.model.RealmUserModel
